backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
pymongo>=4.7.0
pandas>=2.2.0
numpy>=1.26.0
openai>=1.55.0
httpx<0.28.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
